Extract frontend page state hooks
