+++ ProjectPortfolio2026/projectportfolio2026.client/src/features/projects/ProjectDetailPage.tsx
@@ -1,13 +1,6 @@
-import { startTransition, useEffect, useState } from 'react';
 import profilePlaceholder from '../../assets/Placeholders/Profile-Placeholder.png';
 import projectImageUnavailable from '../../assets/Placeholders/Project-Image-Unavailable.png';
-import {
-    fetchResponsePayloadWithStartupRetry,
-    RetryableApiError,
-    startupRetryMessage
-} from '../../app/api';
 import type { NavigateFn } from '../../app/navigation';
-import type { ProjectDetail } from '../../app/types';
 import {
     buildProjectsPath,
     formatFullDate,
@@ -16,6 +9,7 @@ import {
 } from '../../appSupport';
 import { InternalLink } from '../../components/common/InternalLink';
 import { MediaFrame } from '../../components/common/MediaFrame';
+import { useProjectDetail } from '../../hooks/useProjectDetail';
 import { ScreenshotCarousel } from '../media/ScreenshotCarousel';
 
 interface ProjectDetailPageProps {
@@ -29,64 +23,7 @@ export function ProjectDetailPage({
     listSearch,
     onNavigate
 }: ProjectDetailPageProps) {
-    const [project, setProject] = useState<ProjectDetail | null>(null);
-    const [isLoading, setIsLoading] = useState(true);
-    const [error, setError] = useState<string | null>(null);
-    const [isMissing, setIsMissing] = useState(false);
-
-    useEffect(() => {
-        const requestId = crypto.randomUUID();
-        const controller = new AbortController();
-
-        setIsLoading(true);
-        setError(null);
-        setIsMissing(false);
-
-        void loadProject();
-
-        return () => controller.abort();
-
-        async function loadProject() {
-            try {
-                const query = new URLSearchParams({ requestId });
-                const { response, payload } = await fetchResponsePayloadWithStartupRetry<ProjectDetail>(
-                    `/api/projects/${projectId}?${query.toString()}`,
-                    {
-                        signal: controller.signal
-                    },
-                    'Unable to load this project right now.',
-                    [404]
-                );
-
-                if (response.status === 404) {
-                    setIsMissing(true);
-                    setProject(null);
-                    return;
-                }
-
-                if (payload === null) {
-                    throw new RetryableApiError(startupRetryMessage);
-                }
-
-                const projectResponse = payload as ProjectDetail;
-
-                if (projectResponse.requestId !== requestId) {
-                    return;
-                }
-
-                startTransition(() => setProject(projectResponse));
-            } catch (caughtError) {
-                if ((caughtError as Error).name === 'AbortError') {
-                    return;
-                }
-
-                setError(caughtError instanceof Error ? caughtError.message : 'Unable to load this project right now.');
-                setProject(null);
-            } finally {
-                setIsLoading(false);
-            }
-        }
-    }, [projectId]);
+    const { project, isLoading, error, isMissing } = useProjectDetail(projectId);
 
     const backPath = buildProjectsPath(listSearch);
