+++ ProjectPortfolio2026/projectportfolio2026.client/src/features/projects/ProjectListPage.tsx
@@ -1,22 +1,16 @@
-import { Fragment, startTransition, useDeferredValue, useEffect, useRef, useState } from 'react';
+import { Fragment } from 'react';
 import projectImageUnavailable from '../../assets/Placeholders/Project-Image-Unavailable.png';
-import { fetchJsonWithStartupRetry } from '../../app/api';
 import type { NavigateFn } from '../../app/navigation';
-import type { ProjectListResponse, ProjectSummary } from '../../app/types';
+import type { ProjectSummary } from '../../app/types';
 import {
     buildDetailPath,
-    buildListSearch,
-    buildProjectsPath,
-    createRouteKey,
     formatProjectDates,
     getProjectYearSpacerLabel,
-    mergeProjects,
     type ListFilters
 } from '../../appSupport';
 import { InternalLink } from '../../components/common/InternalLink';
 import { MediaFrame } from '../../components/common/MediaFrame';
-
-const pageSize = 6;
+import { useProjectList } from '../../hooks/useProjectList';
 
 interface ProjectListPageProps {
     filters: ListFilters;
@@ -27,193 +21,29 @@ export function ProjectListPage({
     filters,
     onNavigate
 }: ProjectListPageProps) {
-    const [searchInput, setSearchInput] = useState(filters.searchInput);
-    const deferredSearch = useDeferredValue(searchInput.trim());
-    const [selectedSkills, setSelectedSkills] = useState<string[]>(filters.selectedSkills);
-    const [projects, setProjects] = useState<ProjectSummary[]>([]);
-    const [availableSkills, setAvailableSkills] = useState<string[]>([]);
-    const [page, setPage] = useState(1);
-    const [totalCount, setTotalCount] = useState(0);
-    const [hasMore, setHasMore] = useState(false);
-    const [isInitialLoad, setIsInitialLoad] = useState(true);
-    const [isLoading, setIsLoading] = useState(false);
-    const [error, setError] = useState<string | null>(null);
-    const latestRequestIdRef = useRef<string | null>(null);
-    const sentinelRef = useRef<HTMLDivElement | null>(null);
-    const previousQueryKeyRef = useRef<string | null>(null);
-    const routeKey = createRouteKey(filters);
-    const previousRouteKeyRef = useRef(routeKey);
-    const fetchQueryKey = createRouteKey({
-        searchInput: deferredSearch,
-        selectedSkills
-    });
-    const listSearch = buildListSearch({
+    const {
         searchInput,
-        selectedSkills
+        selectedSkills,
+        projects,
+        availableSkills,
+        totalCount,
+        hasMore,
+        isInitialLoad,
+        isLoading,
+        error,
+        listSearch,
+        sentinelRef,
+        handleSearchChange,
+        toggleSkill,
+        clearFilters
+    } = useProjectList({
+        filters,
+        onNavigate
     });
-
-    useEffect(() => {
-        if (previousRouteKeyRef.current === routeKey) {
-            return;
-        }
-
-        previousRouteKeyRef.current = routeKey;
-        setSearchInput(filters.searchInput);
-        setSelectedSkills(filters.selectedSkills);
-        setPage(1);
-        setProjects([]);
-        setAvailableSkills([]);
-        setTotalCount(0);
-        setHasMore(false);
-        setError(null);
-        setIsInitialLoad(true);
-    }, [filters.searchInput, filters.selectedSkills, routeKey]);
-
-    useEffect(() => {
-        const nextPath = buildProjectsPath(listSearch);
-        const currentPath = `${window.location.pathname}${window.location.search}`;
-
-        if (currentPath !== nextPath) {
-            onNavigate(nextPath, { replace: true, preserveScroll: true });
-        }
-    }, [listSearch, onNavigate]);
-
-    useEffect(() => {
-        const requestId = crypto.randomUUID();
-        const controller = new AbortController();
-        const isLoadingMore = page > 1 && previousQueryKeyRef.current === fetchQueryKey;
-        const queryKeyChanged = previousQueryKeyRef.current !== fetchQueryKey;
-
-        latestRequestIdRef.current = requestId;
-        setIsLoading(true);
-        setError(null);
-
-        void loadProjects();
-
-        return () => controller.abort();
-
-        async function loadProjects() {
-            try {
-                const responses = queryKeyChanged
-                    ? await loadPages(1, page)
-                    : [await loadSinglePage(page)];
-
-                if (latestRequestIdRef.current !== requestId || responses.length === 0) {
-                    return;
-                }
-
-                startTransition(() => {
-                    const latestResponse = responses[responses.length - 1];
-                    const incomingProjects = responses.flatMap(response => response.items);
-
-                    setProjects(currentProjects => isLoadingMore
-                        ? mergeProjects(currentProjects, incomingProjects)
-                        : mergeProjects([], incomingProjects));
-                    setAvailableSkills(latestResponse.availableSkills);
-                    setTotalCount(latestResponse.totalCount);
-                    setHasMore(latestResponse.hasMore);
-                });
-
-                previousQueryKeyRef.current = fetchQueryKey;
-            } catch (caughtError) {
-                if ((caughtError as Error).name === 'AbortError' || latestRequestIdRef.current !== requestId) {
-                    return;
-                }
-
-                setError(caughtError instanceof Error ? caughtError.message : 'Unable to load projects right now.');
-                if (!isLoadingMore) {
-                    setProjects([]);
-                    setAvailableSkills([]);
-                    setTotalCount(0);
-                    setHasMore(false);
-                }
-            } finally {
-                if (latestRequestIdRef.current === requestId) {
-                    setIsInitialLoad(false);
-                    setIsLoading(false);
-                }
-            }
-        }
-
-        async function loadPages(startPage: number, endPage: number) {
-            const pages: ProjectListResponse[] = [];
-
-            for (let currentPage = startPage; currentPage <= endPage; currentPage += 1) {
-                const response = await loadSinglePage(currentPage);
-                pages.push(response);
-            }
-
-            return pages;
-        }
-
-        async function loadSinglePage(currentPage: number) {
-            const query = new URLSearchParams({
-                page: currentPage.toString(),
-                pageSize: pageSize.toString(),
-                requestId
-            });
-
-            if (deferredSearch.length > 0) {
-                query.set('search', deferredSearch);
-            }
-
-            if (selectedSkills.length > 0) {
-                query.set('skills', selectedSkills.join(','));
-            }
-
-            const listResponse = await fetchJsonWithStartupRetry<ProjectListResponse>(
-                `/api/projects?${query.toString()}`,
-                {
-                    signal: controller.signal
-                },
-                'Unable to load projects right now.'
-            );
-
-            if (listResponse.requestId !== requestId) {
-                throw new DOMException('Stale response.', 'AbortError');
-            }
-
-            return listResponse;
-        }
-    }, [deferredSearch, fetchQueryKey, page, selectedSkills]);
-
-    useEffect(() => {
-        const sentinel = sentinelRef.current;
-        if (!sentinel || !hasMore || isLoading) {
-            return;
-        }
-
-        const observer = new IntersectionObserver(entries => {
-            if (entries.some(entry => entry.isIntersecting)) {
-                setPage(currentPage => currentPage + 1);
-            }
-        }, { rootMargin: '240px 0px' });
-
-        observer.observe(sentinel);
-        return () => observer.disconnect();
-    }, [hasMore, isLoading, projects.length]);
-
-    function handleSearchChange(nextValue: string) {
-        setSearchInput(nextValue);
-        setPage(1);
-    }
-
-    function toggleSkill(skill: string) {
-        setSelectedSkills(currentSkills =>
-            currentSkills.includes(skill)
-                ? currentSkills.filter(currentSkill => currentSkill !== skill)
-                : [...currentSkills, skill]);
-        setPage(1);
-    }
-
-    function clearFilters() {
-        setSearchInput('');
-        setSelectedSkills([]);
-        setPage(1);
-    }
 
     const showingCount = projects.length;
-    const isEmpty = !isLoading && !error && projects.length === 0;
+    const activeFilterCount = selectedSkills.length + (searchInput.trim().length > 0 ? 1 : 0);
+    const isEmpty = !isLoading && !error && showingCount === 0;
 
     return (
         <main className="portfolio-page">
@@ -239,7 +69,7 @@ export function ProjectListPage({
                         </div>
                         <div className="stat-card">
                             <span className="stat-label">Active Filters</span>
-                            <strong>{selectedSkills.length + (deferredSearch.length > 0 ? 1 : 0)}</strong>
+                            <strong>{activeFilterCount}</strong>
                         </div>
                     </div>
                 </div>
