+++ ProjectPortfolio2026/projectportfolio2026.client/src/hooks/useMediaQuery.ts
@@ -1,0 +1,18 @@
+import { useEffect, useState } from 'react';
+
+export function useMediaQuery(query: string) {
+    const [matches, setMatches] = useState(() => window.matchMedia(query).matches);
+
+    useEffect(() => {
+        const mediaQuery = window.matchMedia(query);
+        const updateMatches = (event?: MediaQueryListEvent) => {
+            setMatches(event?.matches ?? mediaQuery.matches);
+        };
+
+        updateMatches();
+        mediaQuery.addEventListener('change', updateMatches);
+        return () => mediaQuery.removeEventListener('change', updateMatches);
+    }, [query]);
+
+    return matches;
+}
